Update for README.md
When i first Navigated the repo, i was searching for the link to View site. I could'nt find it and finally i googled the site and found it. I think it would be easier for newcomers and general github users to just navigate the site with just a click from readme, rather than googling it.